requirement.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn-pandas==1.5.0
threadpoolctl==3.1.0
zipp==3.14.0
